move defaultEventParametersForInit and funcs to functions file
