+++ packages/analytics/src/functions.ts
@@ -23,6 +23,12 @@ import {
 } from './public-types';
 import { Gtag } from './types';
 import { GtagCommand } from './constants';
+
+/**
+ * Event parameters to set on 'gtag' during initialization.
+ */
+export let defaultEventParametersForInit: CustomParams | undefined;
+
 /**
  * Logs an analytics event through the Firebase SDK.
  *
@@ -141,4 +147,23 @@ export async function setAnalyticsCollectionEnabled(
 ): Promise<void> {
   const measurementId = await initializationPromise;
   window[`ga-disable-${measurementId}`] = !enabled;
+}
+
+/**
+ * Sets the variable {@link defaultEventParametersForInit} for use in the initialization of
+ * analytics.
+ *
+ * @param customParams Any custom params the user may pass to gtag.js.
+ */
+export function _setDefaultEventParametersForInit(
+  customParams: CustomParams
+): void {
+  if (defaultEventParametersForInit) {
+    defaultEventParametersForInit = {
+      ...defaultEventParametersForInit,
+      ...customParams
+    };
+  } else {
+    defaultEventParametersForInit = customParams;
+  }
 }
